requirements.txt
PyQt6>=6.6.0
openpyxl>=3.1.0
python-docx>=1.1.0
docxtpl>=0.16.0
xlwings>=0.30.0
Pillow>=10.0.0
pywin32>=305
